+++ test/test_parser.cpp
@@ -2668,7 +2668,8 @@ TEST(Parser_OperatorIN)
 
 // we won't support full object comparisons until we have stable keys in core, but as an exception
 // we allow comparison with null objects because we can serialise that and bindings use it to check agains nulls.
-TEST(Parser_RowIndex) {
+TEST(Parser_RowIndex)
+{
     Group g;
     TableRef table = g.add_table("table");
     size_t int_col_ndx = table->add_column(type_Int, "ints", true);
@@ -2692,6 +2693,29 @@ TEST(Parser_RowIndex) {
     CHECK_EQUAL(q1.count(), 2);
 
     CHECK_THROW_ANY(verify_query(test_context, table, "link == link", 3));
+}
+
+TEST(Parser_Between)
+{
+    Group g;
+    TableRef table = g.add_table("table");
+    size_t int_col_ndx = table->add_column(type_Int, "age", true);
+    size_t between_col_ndx = table->add_column(type_Int, "between", true);
+    table->add_empty_row(3);
+    for (size_t i = 0; i < table->size(); ++i) {
+        table->set_int(int_col_ndx, i, i + 24);
+        table->set_int(between_col_ndx, i, i);
+    }
+
+    // normal querying on a property named "between" is allowed.
+    verify_query(test_context, table, "between == 0", 1);
+    verify_query(test_context, table, "between > 0", 2);
+    verify_query(test_context, table, "between <= 3", 3);
+
+    // operator between is not supported yet, but we at least use a friendly error message.
+    std::string message;
+    CHECK_THROW_ANY_GET_MESSAGE(verify_query(test_context, table, "age between {20, 25}", 1), message);
+    CHECK(message.find("Invalid Predicate. The 'between' operator is not supported yet, please rewrite the expression using '>' and '<'.") != std::string::npos);
 }
 
 #endif // TEST_PARSER